change "Deleted" message to "Filtered"
Commands like "search" are non-destructive but their outputs suggest
that goroutines are being deleted. Use the term "filtered" instead of
"deleted", which can serve reasonably well for both destructive and
non-destructive commands.

This message is still a little nonsensical for some commands and we
could probably pass a format string from the caller instead of having
it owned by `withCondition`.