remove writes to stdout from the functional components.
First pass at injecting an `io.Writer` into the functional components so they're
not making direct writes to stdout. This will let us test the expected outputs
without running the REPL.

Add a dependency on `mattn/colorable` so we have a wrapper that remove the ANSI
term sequences as needed, and provide better support for Windows in the process.